fix: Use sudo to change default shell in zsh task
The `chsh` command requires root privileges to change the default shell for the current user. This commit updates the `set-zsh-default-shell` task in `taskfile.tools.yaml` to use `sudo` when executing the `chsh` command. This ensures that the command has the necessary permissions to modify the default shell. Also explicitly specify the user to change the shell for using $(whoami).