feat: Add interactive SSH configuration creation to taskfile
This commit enhances the `taskfile.git.yaml` by introducing an interactive prompt that allows users to create an SSH configuration for their git provider. If the user opts to create the configuration, they can input their SSH public key, which is then added to the configuration file, streamlining the setup process for SSH key management.